+++ libkbfs/root_metadata_test.go
@@ -473,71 +473,6 @@ func TestBareRootMetadataUnknownFields(t *testing.T) {
 	testStructUnknownFields(t, makeFakeBareRootMetadataFuture(t))
 }
 
-// TODO: Have MDV3 version.
-func TestIsValidRekeyRequestBasicV2(t *testing.T) {
-	config := MakeTestConfigOrBust(t, "alice")
-	defer config.Shutdown()
-
-	// Sign the writer metadata
-	tlfID := FakeTlfID(1, false)
-
-	h := parseTlfHandleOrBust(t, config, "alice", false)
-	brmd, err := MakeInitialBareRootMetadataV2(
-		tlfID, h.ToBareHandleOrBust())
-	require.NoError(t, err)
-
-	err = brmd.SignWriterMetadataInternally(
-		context.Background(), config.Codec(), config.Crypto())
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Copy bit unset.
-	newBrmd, err := MakeInitialBareRootMetadataV2(
-		tlfID, h.ToBareHandleOrBust())
-	require.NoError(t, err)
-	ok, err := newBrmd.IsValidRekeyRequest(
-		config.Codec(), brmd, newBrmd.LastModifyingWriter(), nil, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if ok {
-		t.Fatal("Expected invalid rekey request due to unset copy bit")
-	}
-
-	// Set the copy bit; note the writer metadata is the same.
-	newBrmd.SetWriterMetadataCopiedBit()
-
-	// Writer metadata siginfo mismatch.
-	config2 := MakeTestConfigOrBust(t, "bob")
-	defer config2.Shutdown()
-
-	err = newBrmd.SignWriterMetadataInternally(
-		context.Background(), config2.Codec(), config2.Crypto())
-	if err != nil {
-		t.Fatal(err)
-	}
-	ok, err = newBrmd.IsValidRekeyRequest(
-		config.Codec(), brmd, newBrmd.LastModifyingWriter(), nil, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if ok {
-		t.Fatal("Expected invalid rekey request due to mismatched writer metadata siginfo")
-	}
-
-	// Replace with copied signature.
-	newBrmd.WriterMetadataSigInfo = brmd.WriterMetadataSigInfo
-	ok, err = newBrmd.IsValidRekeyRequest(
-		config.Codec(), brmd, newBrmd.LastModifyingWriter(), nil, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if !ok {
-		t.Fatal("Expected valid rekey request")
-	}
-}
-
 func TestMakeRekeyReadError(t *testing.T) {
 	config := MakeTestConfigOrBust(t, "alice", "bob")
 	defer config.Shutdown()
